profile wip

+++ SystemyWP.Data/Models/LegalAppModels/Clients/LegalAppClient.cs
@@ -5,7 +5,7 @@ using SystemyWP.Data.Models.LegalAppModels.Cases;
 
 namespace SystemyWP.Data.Models.LegalAppModels.Clients
 {
-    public class LegalAppClient : LegalAppProtectedDataBaseModel<int>
+    public class LegalAppClient : LegalAppProtectedDataBaseModel<long>
     {
         public bool Active { get; set; }
